Started using the Gauss-Seidel method from PMGC
I started implementing a small multigrid solver in the fashion of pmgc.
Unfortunately, I did not manage to get the Gauss-Seidel method converge
on its own. In order to fix this I went back to my own Gauss-Seidel
method, which I know works. It is now possible to compare the two
methods and find the error.