Found the reason why the pmgc version diverged
It turns out that omgc has a a different index ordering than my version.
For them, the z-direction is the fastest moving as it is the case in
fortran. The is usually not a problem, but it becomes a problem if you
introduce maps that are shifted aling one direction, such as
epsilon_maps. In order to fix this issue, one should always exchange the
epsilonz_map with the epsilon_x map when working with pmgc