Added a intermediate abstraction layer for the Domain class
The Grid class had the problem, that it held the values in value_buff,
which is device allocated memory, which was never freed in the
destructor of the Grid class. At the same time, it was not possible to
to create a shared_ptr for  each object of the class, as this way the
object would no longer be able to be copied to the device and be used as
functors inside the kernel.

I, therefore, introduced a Grid_wrapper class, which directly inherits
from the Grid class which has the additional shared_ptr. The old code
remains intact this way, due to the slicing mechanism